+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "cmake.sourceDirectory": "D:/projects/my_store/linux"
+}
