+++ lib/my_store_app.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+class MyStoreApp extends StatelessWidget {
+  const MyStoreApp({super.key});
+
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: Scaffold(
+        appBar: AppBar(
+          backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+          title: const Text('MyStore'),
+        ),
+        body: const Center(
+          child: Text('MyStore'),
+        ),
+      ),
+    );
+  }
+}
